+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.41.0
 
-- Removed `options.mapper` in favor of `rename` which takes an object of keys to be renamed.
+- Removed `options.mapper` in favor of `options.rename` which takes an object of keys to be renamed.
+The `rename` option is available for syncing and converting a schema to a table.
 
 # 0.40.0